requirements.txt
webdriver-manager==3.5.3
behave==1.2.6
selenium==4.1.2
PyHamcrest==2.0.3